final version of SHAP Analysis with globabl view
